requirements.txt
flask==3.0.0
pandas==2.2.3
scikit-learn==1.5.2
numpy>=1.26.0
werkzeug==3.0.1
scipy>=1.11.4
joblib==1.3.2
gunicorn>=21.2.0
pdfminer.six>=20221105
python-docx>=0.8.11
docx2txt>=0.8
PyPDF2>=3.0.1
Flask-Cors==4.0.0
pymongo[srv]==4.6.3
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.2
requests==2.32.3
